\name{Rdummy-package}
\alias{Rdummy-package}
\alias{Rdummy}
\docType{package}
\title{
\packageTitle{Rdummy}
}
\description{
\packageDescription{Rdummy}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Rdummy}

Maintainer: \packageMaintainer{Rdummy}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Rdummy)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Rdummy}}
%\section{Documentation Index}{\packageIndices{Rdummy}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
